Adds BLOG to gitmol.com
These BLOG entries are the ones that were posted on
Manning's forums, which are now defunct. These posts
exist in the Livebook pages, but as archived posts,
that isn't really laid out. I resuscitated the content,
laying out the posts using the static site generator Hexo
(and a basic Bootstrap theme).

https://hexo.io/
https://github.com/cgmartin/hexo-theme-bootstrap-blog